combine misc.c and process.c
